backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data processing and ML
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# RSS parsing
feedparser==6.0.10

# HTTP client for API calls
requests==2.31.0
httpx==0.25.0

# Background tasks and scheduling
apscheduler==3.10.4

# Data validation
pydantic==2.4.2

# Environment variables
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Development dependencies
pytest==7.4.2
pytest-asyncio==0.21.1
black==23.9.1
flake8==6.1.0

# Optional: if you need async database operations
aiosqlite==0.19.0 